+++ lib/simplecov/result.rb
@@ -102,7 +102,7 @@ module SimpleCov
       formatter = SimpleCov.formatter
       return nil if formatter.nil?
 
-      formatter.new.format(self)
+      Formatter.instance_for(formatter).format(self)
     end
 
     # Defines when this result has been created. Defaults to Time.now